--- a/.gitmodules
+++ b/.gitmodules
@@ -1,0 +1,3 @@
+[submodule "benchmark/git"]
+	path = benchmark/git
+	url = https://github.com/git/git.git