--- a/hs-git-gsub.cabal
+++ b/hs-git-gsub.cabal
@@ -67,3 +67,17 @@ test-suite hs-git-gsub-test
   default-language:   Haskell2010
   build-tool-depends: hspec-discover:hspec-discover
   hs-source-dirs:     test
+
+benchmark hs-git-gsub-benchmark
+  type:               exitcode-stdio-1.0
+  main-is:            Main.hs
+  build-depends:
+      base
+    , directory
+    , process
+    , criterion
+    , hs-git-gsub
+
+  hs-source-dirs:     benchmark
+  default-language:   Haskell2010
+  ghc-options:        -threaded -rtsopts -with-rtsopts=-N -Wall