Benchmark result may not trustworthy, but it seems to be some actual improvements. 

With strict ByteString(cf7c55c661f0c712638715a18f35d9cacca2a895)

  | lower bound | estimate | upper bound
-- | -- | -- | --
OLS regression | 1.30 s | 10.6 s | 23.1 s
R² goodness-of-fit | 0.715 | 0.865 | 1.00
Mean execution time | 10.6 s | 13.4 s | 17.3 s
Standard deviation | 1.37 s | 3.70 s | 4.90 s


With lazy ByteString(c1ba2bc737950de5eba71fea9126eacb4a5f9fa8)

  | lower bound | estimate | upper bound
-- | -- | -- | --
OLS regression | 1.50 s | 1.74 s | 1.98 s
R² goodness-of-fit | 0.991 | 0.997 | 1.00
Mean execution time | 1.73 s | 1.77 s | 1.80 s
Standard deviation | 34.6 ms | 42.8 ms | 48.3 ms